Ticket #1613: Test west build with no docker for HIL

+++ samples/embedded/platform_dep/zephyr/ci/build-zephyr-samples.sh
@@ -19,21 +19,22 @@
 
 
 set -e
-export SCRIPT=build-zephyr-samples-docker.sh
+export SCRIPT="$(pwd)/build-zephyr-samples-west.sh"
 cd "$(dirname "$0")"/../../../../..
 HOST_DIR=$(pwd)
 
 function usage() {
     echo "By default builds a predefined set of Zephyr applications"
     echo "To build a specific Zephyr applications use -b <BOARD> -a <APPLICATION>"
-    echo "Usage : $0 [-h] [-i] [-a <APPLICATION>] [-b <BOARD>] [--ip <IP_ADDRESS>] [--log <PATH>] [--bin <PATH>]"
+    echo "Usage : $0 [-h] [-i] [-a <APPLICATION>] [-b <BOARD>] [--ip <IP_ADDRESS>] [--log <PATH>] [--bin <PATH>] [--docker]"
     echo "  $0                      : build predefined application/boards"
     echo "  $0 -i                   : Start an interactive session in zephyr docker. see ${SCRIPT} for more help"
     echo "  $0 -b                   : Allows user to specify a <BOARD>"
     echo "  $0 -a                   : Allows user to specify an <APPLICATION>"
     echo "  $0 --ip <IP_ADDRESS>    : Configure IP Adress of ethernet interface"
-    echo "--log <PATH> : give a specific path to store the logs"
-    echo "--bin <PATH> : give a specific path to store the bins"
+    echo "--log <PATH>              : give a specific path to store the logs"
+    echo "--bin <PATH>              : give a specific path to store the bins"
+    echo "--docker                  : run the build script in docker"
     echo "  $0 -h                   : This help"
     exit 0
 }
@@ -45,6 +46,7 @@ GET_BOARD=
 GET_APP=
 LOG_PATH=
 BIN_PATH=
+DOCKER=false
 
 while [[ ! -z $1 ]]; do
     PARAM=$1
@@ -56,6 +58,7 @@ while [[ ! -z $1 ]]; do
     [[ ${PARAM-} == "--ip" ]] && OPT_IP_ADDRESS="--ip $1" && shift && continue
     [[ ${PARAM-} == "--log" ]] && LOG_PATH="--log $1" && shift && continue
     [[ ${PARAM-} == "--bin" ]] && BIN_PATH="--bin $1" && shift && continue
+    [[ ${PARAM-} == "--docker" ]] && DOCKER=true && continue
 
     echo "Unknown parameter : ${PARAM}"
     exit 2
@@ -64,13 +67,16 @@ done
 [[ ! -z $LOG_PATH ]] && LOG_PATH="--log ${LOG_PATH}"
 [[ ! -z $BIN_PATH ]] && BIN_PATH="--bin ${BIN_PATH}"
 
-source .docker-images.sh
-docker inspect ${ZEPHYR_DIGEST} 2>/dev/null >/dev/null  || fail "Docker image not installed: ${ZEPHYR_DIGEST}"
 
-echo "Mapping ${HOST_DIR} to DOCKER '/workdir'"
-$IS_INTERACTIVE && echo "Running an interactive session on ${ZEPHYR_DIGEST}" && \
-    (docker run -it --rm -v ${HOST_DIR}:/zephyrproject/modules/lib/s2opc -w /zephyrproject/modules/lib/s2opc ${ZEPHYR_DIGEST})
-$IS_INTERACTIVE && exit 1
+if [ "$DOCKER" = true ]; then
+    source .docker-images.sh
+    docker inspect ${ZEPHYR_DIGEST} 2>/dev/null >/dev/null  || fail "Docker image not installed: ${ZEPHYR_DIGEST}"
+
+    echo "Mapping ${HOST_DIR} to DOCKER '/workdir'"
+    $IS_INTERACTIVE && echo "Running an interactive session on ${ZEPHYR_DIGEST}" && \
+        (docker run -it --rm -v ${HOST_DIR}:/zephyrproject/modules/lib/s2opc -w /zephyrproject/modules/lib/s2opc ${ZEPHYR_DIGEST})
+    $IS_INTERACTIVE && exit 1
+fi
 
 function build() {
   export BOARD=$1
@@ -78,13 +84,20 @@ function build() {
   export APP=$1
   shift
   export ADD_CONF=$*
+  
   echo "Starting docker to build ${APP} for ${BOARD}"
-
+  
+  if [ "$DOCKER" = true ]; then
   (docker run --rm -v ${HOST_DIR}:/zephyrproject/modules/lib/s2opc -w /zephyrproject/modules/lib/s2opc ${ZEPHYR_DIGEST}\
-     samples/embedded/platform_dep/zephyr/ci/${SCRIPT} ${BOARD} ${APP} ${OPT_IP_ADDRESS} ${LOG_PATH} ${BIN_PATH} ${ADD_CONF})&
-  wait $!
+    samples/embedded/platform_dep/zephyr/ci/${SCRIPT} ${BOARD} ${APP} ${OPT_IP_ADDRESS} ${LOG_PATH} ${BIN_PATH} ${ADD_CONF})& 
+    wait $!
+  else
+    (${SCRIPT} ${BOARD} ${APP} ${OPT_IP_ADDRESS} ${LOG_PATH} ${BIN_PATH} --nodocker ${ADD_CONF})& wait $!
+  fi
   echo "Result = $?"
 }
+
+
 if [ -z "${GET_BOARD}" ] && [ -z "${GET_APP}" ]; then
     rm -rf build_zephyr/* 2>/dev/null
     mkdir -p build_zephyr
